// supabase/functions/_shared/database.types.ts
export interface Database {
// ...
  public: {
    Tables: {
      learning_plan_items: {
        Row: {
          details: string
          id: number
          learning_plan_id: number
          title: string
        }
        Insert: {
          details: string
          id?: never
          learning_plan_id: number
          title: string
        }
        Update: {
          details?: string
          id?: never
          learning_plan_id?: number
          title?: string
        }
        Relationships: [
          {
            foreignKeyName: "learning_plan_items_learning_plan_id_fkey"
            columns: ["learning_plan_id"]
            referencedRelation: "learning_plans"
            referencedColumns: ["id"]
          }
        ]
      }
      learning_plans: {
        Row: {
          id: number
          lesson_duration: string
          number_of_lessons: number
          target_language: string
          target_language_level: string
        }
        Insert: {
          id?: never
          lesson_duration: string
          number_of_lessons: number
          target_language: string
          target_language_level: string
        }
        Update: {
          id?: never
          lesson_duration?: string
          number_of_lessons?: number
          target_language?: string
          target_language_level?: string
        }
        Relationships: []
      }
      lessons: {
        Row: {
          id: number
          learning_plan_item_id: number
          transcript: string
        }
        Insert: {
          id?: never
          learning_plan_item_id: number
          transcript: string
        }
        Update: {
          id?: never
          learning_plan_item_id?: number
          transcript?: string
        }
        Relationships: [
          {
            foreignKeyName: "lessons_learning_plan_item_id_fkey"
            columns: ["learning_plan_item_id"]
            referencedRelation: "learning_plan_items"
            referencedColumns: ["id"]
          }
        ]
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
// ...
}